removed comments and fixed the drag and drop issue
