+++ Desktop/src/parser/utils/layout.ts
@@ -1,4 +1,6 @@
-import { isNode, Node } from 'reactflow';
+import { useEffect, FunctionComponent } from 'react';
+import { isNode, Node, useReactFlow, useStore, useNodes, } from 'reactflow';
+
 
 const groupNodesByColumn = (elements: Node[]) => {
     return elements.reduce((elementsGrouped: any, element: Node) => {
@@ -60,3 +62,24 @@ const groupNodesByColumn = (elements: Node[]) => {
     return newElements.nodes;
   
   };
+
+export const AutoLayout: FunctionComponent<AutoLayoutProps> = () => {
+  const { fitView } = useReactFlow();
+  const nodes = useNodes();
+  const setNodes = useStore(state => state.setNodes);
+
+  useEffect(() => {
+    if (nodes.length === 0 || !nodes[0].width) {
+      return;
+    }
+
+    const nodesWithOrginalPosition = nodes.filter(node => node.position.x === 0 && node.position.y === 0);
+    if (nodesWithOrginalPosition.length > 1) {
+      const calculatedNodes = calculateNodesForDynamicLayout(nodes);
+      setNodes(calculatedNodes);
+      fitView();
+    }
+  }, [nodes]);
+
+  return null;
+};
