+++ Desktop/src/renderer/AddNodesDnD/AddButton.tsx
@@ -60,11 +60,11 @@ function DropdownMenu({ nodes, setNodes }) {
 
     function DropdownItem(props) {
         return (
-            <a href="#" className="menu-item" onClick={() => props.goToMenu && setActiveMenu(props.goToMenu)}>
+            <div className="menu-item" onClick={() => props.goToMenu && setActiveMenu(props.goToMenu)}>
                 <span className="icon-button">{props.leftIcon}</span>
                 {props.children}
                 <span className="icon-right">{props.rightIcon}</span>
-            </a>
+            </div>
         );
     }
