+++ Desktop/src/renderer/containers/Workbench/index.tsx
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import ScenarioEditor from './components/ScenarioEditor';
 import ScenarioVisualizer from './components/ScenarioVisualizer';
 import SideBar from './components/SideBar';
